services/0vers33r/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.9.0
cryptography==45.0.3
deprecation==2.1.0
firebase-admin==6.9.0
firebase-functions==0.1.2
Flask==3.1.1
flask-cors==6.0.0
functions-framework==3.8.3
google-api-core
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-events
google-genai==1.23.0
google-resumable-media==2.7.2
googleapis-common-protos
grpcio==1.72.1
grpcio-status
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
yara-python==4.5.4
# OSS platform dependencies
minio==7.2.5
psycopg2-binary==2.9.9
nats-py==2.6.0
google-generativeai==0.8.3
